DIT-409 PR comments and refactoring test data to live within test folder.
