+++ build.sbt
@@ -31,14 +31,20 @@ lazy val microservice = (project in file("."))
   )
   .settings(inConfig(TemplateTest)(Defaults.testSettings): _*)
   .settings(
-    unmanagedSourceDirectories in Test := Seq((baseDirectory in Test).value / "test/unit"),
+    unmanagedSourceDirectories in Test := Seq(
+      (baseDirectory in Test).value / "test/unit",
+      (baseDirectory in Test).value / "test/util"
+    ),
     addTestReportOption(Test, "test-reports")
   )
   .configs(IntegrationTest)
   .settings(inConfig(TemplateItTest)(Defaults.itSettings): _*)
   .settings(
     Keys.fork in IntegrationTest := false,
-    unmanagedSourceDirectories in IntegrationTest := Seq((baseDirectory in IntegrationTest).value / "test/it"),
+    unmanagedSourceDirectories in IntegrationTest := Seq(
+      (baseDirectory in IntegrationTest).value / "test/it",
+      (baseDirectory in Test).value / "test/util"
+    ),
     addTestReportOption(IntegrationTest, "int-test-reports"),
     testGrouping in IntegrationTest := oneForkedJvmPerTest((definedTests in IntegrationTest).value),
     parallelExecution in IntegrationTest := false)
